vimrc: FindTabStyle: Optionally pass ignore regex (fixes #1)
Initially, the reason for checking that a line from which the indent
style should be deduced should not match prev_line_regex, is that in
case of C++, class declarations usually start with

1: class T {
2:   public:
3:     int foo;
4: };

Here we do not want to deduce indentation to be 2 spaces (line 2), but
instead we want 4 spaces (line 3).

To fix the issues in #1, explicitly pass the lines to be ignored; e.g.
for C++, do not exclude '{$'.